Color palette for flagged records

flag_colors is a named character vector defining the default colors used to plot occurrence records flagged with mapview_here().

Usage

flag_colors

Format

A named character vector where:

names

Flag labels corresponding to categories generated by the various flag_* and checking functions.

values

Hex color codes or standard R color names used for plotting.

See also

Examples

data(flag_colors)

# View all flag categories and their colors
flag_colors
#>                 Wrong country                   Wrong state 
#>                     "#E69F00"                     "#56B4E9" 
#>                    Cultivated                        Fossil 
#>                     "#16FF32"                       "black" 
#>                   Inaturalist               Outside florabr 
#>                     "#F0E442"                     "#782AB6" 
#>               Outside faunabr                  Outside WCVP 
#>                     "#782AB6"                     "#D55E00" 
#>                  Outside IUCN                  Outside BIEN 
#>                     "#CC79A7"                     "#0072B2" 
#>                    Duplicated                    No flagged 
#>                     "#FFFF80"                     "#009E73" 
#>              Invalid lat/long                Equal lat/long 
#>                      "gray22"                     "#F7E1A0" 
#>                 Zero lat/long              Capital centroid 
#>                     "#85660d"                         "red" 
#>                   Out of year     Country/Province centroid 
#>                      "gray10"                     "orange3" 
#>                      Open sea                    Urban area 
#>                    "darkblue"                        "gray" 
#>                       Outlier             GBIF headquarters 
#>                        "pink"                     "#B00068" 
#>      Biodiversity Institution Artificial Hotspot Occurrence 
#>                   "firebrick"                     "#2ED9FF" 
#>               Geo-Thinned out               Env-Thinned out 
#>                         "red"                        "gold" 
#>              Forcibly removed                     consensus 
#>                      "gray25"                   "chocolate"